Breaking up methods in data import module
This implementation is inflexible, and should be broken
into smaller methods.  As a first start, this commit adds
a 'dry run' feature that dumps the data to stdout rather than
sending it to the Solr index.

Eventually, this will allow users to save import data in the form
it is indexed.  In addition, there will be a more flexible mapping
logic that the one-to-one arrangment here.