Major refactor: dates now ISO8601 timestamps
This is a big set of conditionals designed to get the "date"
field in Solr to serve as the basis for a date_sort field.

For good sorting on the current field (a TextField, not DateTime)
dates of the YYYY-MM-DDTHH:MM:SSZ kind are useful.  Where possible,
the importer transforms dc:date fields into this kind of string, adding
a copy of the original to :pub_date, to be used for display.

TO DO: reconfigure CatalogController's Blacklight configuration to
take advantage of this.  Date sorting should be on :date from now on,
and the date_query_facet_field should be scrutinized to ensure its
behavior is consistent with other features on the site.

More important TO DO: write tests for this!!

Also: remove copyField directive that copies :date to :pub_date, as
you will raise index-time errors if you try to write to a copy field.

Once we are confident we are only putting ISO 8601 compliant timestamps
into the :date field, we can test re-typing that field as DateTime, or
creating a new copy field so typed.