Fixing sort on Author,Title, and Date
These changes go hand-in-hand with the recent changes to the
importer library.  Dates are now converted to something resembling
timestamps, and thus the "date" field in Solr is available for
sorting.  "pub_date" is a related field, where the unmodified string
from the data source is indexed and stored.  Pub date is used for
displaying dates, as well as for the range queries in the current
"Pub Date" facet on the sidebar, as can be seen from the CatController
config.  :pub_date is copied to :date_facet, while :date is copied to
:date_sort

Given that users can sort on date, it seemed appropriate to include
this field in the index view.  Otherwise, index entries will not
reveal much about where they fit in a time period.

Blacklight config also expanded to permit forward- and reverse-sorting
on dates, which seems useful.

As date_sort is copied from the most accurate date field indexed, it
now replaces pub_date_sort as the field used in ordering results.

Solr schema changes involve ensuring :pub_date is available to receive
data at index-time, and this was not possible when it was a copy field.

During testing it became clear why author and title sorting was not
working, and the necessary sort fields have been added to the list of
copy fields, to receive data from their proper sources.

Solr configuration has been updated to add the date field to the
'search' and 'document' handlers.  Note that the latter handler is
used in both building the index and show views.