Prevening truncation of linked field values.
This is a fix of recent issue of subject fields with many entries.
Since the subject_t field is configured with a link to value, the
resulting text contains markup, which in long strings can result in
broken HTML.  Rather than increase the 2000-char setting of the
truncate method call in render_document_show_field_value, we have
added a test to circumvent the truncate method in cases where the
field is flagged as linakble.